Move the location of the asdf_maker subpackage to io.asdf.generator
