+++ dkist/io/asdf/generator/__init__.py
