+++ dkist/io/asdf/generator/helpers.py
