+++ dkist/io/asdf/generator/tests/test_transforms.py
@@ -7,8 +7,8 @@ from astropy.io import fits
 from astropy.time import Time
 import astropy.units as u
 
-from dkist.asdf_maker.helpers import headers_from_filenames
-from dkist.asdf_maker.transforms import (linear_spectral_model, linear_time_model,
+from dkist.io.asdf.generator.helpers import headers_from_filenames
+from dkist.io.asdf.generator.transforms import (linear_spectral_model, linear_time_model,
                                          spatial_model_from_header, spectral_model_from_framewave,
                                          time_model_from_date_obs)
