--- docs/asdfmaker.rst
@@ -1,11 +1,0 @@
-.. _dkist-asdf-maker:
-
-asdf Maker Module
-=================
-
-This module provides tools to make DKIST asdf datasets. It provides tools to
-read DKIST FITS files and generate the gWCS and array structures that are then
-saved in the asdf file.
-
-
-.. automodapi:: dkist.asdf_maker
