+++ docs/io.rst
@@ -24,3 +24,6 @@ API Reference
 
 .. automodapi:: dkist.io.array_containers
    :headings: #^
+
+.. automodapt:: dkist.io.asdf.generator
+   :headings: #^
